Fix partial reads
Partial reads use an internal buffer that is filled and drained into the requestors buffer. If some data is drained we still wait on the file descriptor even though no more data may arrive until the requestor responds (e.g. last chunk of HTTP was drained from the rxbuf here but we still wait for more data though we need to send a response first likely!)